Big and unstable changes.

+++ lib/rwiki/file_utils.rb
@@ -2,6 +2,8 @@ module Rwiki
   module FileUtils
     include ::FileUtils
 
+    PAGE_FILE_EXT = '.txt'
+
     def make_tree(folder_name)
       make_nodes(folder_name)
     end
@@ -30,27 +32,35 @@ module Rwiki
 
     def read_page(page_name)
       return nil unless File.exists?(page_name)
-      return File.read(page_name) { |f| f.read }
+      return File.read(page_name) { |f| f.read } rescue nil
     end
 
     def write_page(page_name, content)
+      return false unless File.exists?(page_name)
+      
       file = File.open(page_name, 'w')
       file.write(content)
       file.close
+
+      return true
+    rescue e
+      return false
     end
 
-    def create_folder(parent_folder_name, base_name)
-      new_folder_name = File.join(parent_folder_name, base_name)
-      mkdir(new_folder_name)
+    def create_folder(parent_folder_name, folder_base_name)
+      folder_name = File.join(parent_folder_name, folder_base_name)
+      return false if File.exists?(folder_name)
+      mkdir(folder_name)
 
-      return new_folder_name
+      return folder_name
     end
 
-    def create_page(parent_folder_name, base_name)
-      new_page_name = File.join(parent_folder_name, base_name) + '.txt'
-      touch(new_page_name)
+    def create_page(parent_folder_name, page_base_name)
+      page_name = File.join(parent_folder_name, page_base_name)
+      return false if File.exists?(page_name)
+      touch(page_name)
 
-      return new_page_name
+      return page_name
     end
 
     def rename_node(old_node_name, new_base_name)
@@ -65,13 +75,15 @@ module Rwiki
     def move_node(node_name, dest_folder_name)
       new_node_name = File.join(dest_folder_name, File.basename(node_name))
       return false if File.exists?(new_node_name)
-      
+
       mv(node_name, dest_folder_name)
       return true, new_node_name
     end
 
     def delete_node(node_name)
+      return false if File.exists?(node_name)
       rm_rf(File.join(node_name))
+      return true
     end
 
   end
